pyproject.toml
[build-system]
requires = [
    "setuptools >= 61.0.0",
    "setuptools_scm >= 7.0.0",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "lymixture"
description = "Package to model lymphatic spread sa mixture of HMMs as implemented in the `lymph-model` package."
authors = [
    {name = "Julian Brönnimann"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "numpy",
    "scipy",
    "pandas",
    "lymph-model >= 1.2",
    # I think these packages should not be depedencies...
    "emcee",
    "lyscripts >= 1.0.0.a0",
]
dynamic = ["version"]

[project.optional-dependencies]
test = [
    "jupyter",
    "pytest",
]

[tool.setuptools_scm]
write_to = "lymixture/_version.py"
local_scheme = "no-local-version"

[tool.setuptools]
packages = ["lymixture"]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "lymixture._version.version"}

[tool.ruff]
exclude = ["*.ipynb"]
